Edited actionPerformed method in the search button listener to use the correct column names in the SQL query.
Added Javadoc comments

Modified code to allow for multiple query selections.
Added checks to see if the department and project lists
are not empty before building the corresponding 'IN' clauses
in the query.

Added a JOIN clause with the department table, allowing the ability
to filter employees based on the selected department names.